refactor: split FHIR module into filters/permissions/pagination/views; move tests
- Extract FHIRLocationFilter to filters.py
- Extract FHIRLocationPermission to permissions.py
- Create FHIRPaginator (LimitOffsetPagination) in pagination.py using _count/_skip
- Rename location.py to views.py; viewset now uses pagination_class
- Register FHIRLocationViewSet in main router (visible in DRF browsable API)
- Move tests to iaso/tests/api/fhir/test_location.py
- Update README: expand FHIR acronym and link to spec
