+++ iaso/api/fhir/urls.py
@@ -5,7 +5,7 @@ URL configuration for FHIR API endpoints
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from .location import FHIRLocationViewSet
+from .views import FHIRLocationViewSet
 
 
 # Create router for FHIR endpoints
